Use Gradle Wrapper to invoke Gradle
The Gradle Wrapper is the preferred way to execute Gradle, because
it ensures that users will have a consistent build environment, and it
conveniently downloads Gradle if the user doesn't have it.

This change configures the Gradle wrapper for Gradle 4.6, which is the
latest.

See https://docs.gradle.org/current/userguide/gradle_wrapper.html